third_party/WebKit/LayoutTests/http/tests/inspector/elements/event-listeners-framework-with-service-worker-expected.txt

Tests framework event listeners output in Sources panel when service worker is present.

Inspect Me
Selecting service worker thread
Context is dedicated worker: true
Dumping listeners
Selecting main thread
Context is dedicated worker: false
Dumping listeners

======== load ========
== normal
[expanded] WindowRemoveevent-listeners-framework-with-service-worker.html:56
    useCapture: false
    handler: function onload(event) {
  runTest()
}
